+++ api/spec/mailer/mno_enterprise/system_notification_mailer_spec.rb
@@ -13,10 +13,10 @@ module MnoEnterprise
     def user_vars(user)
       { first_name: user.name, last_name: user.surname, full_name: "#{user.name} #{user.surname}".strip }
     end
-    
+
     def invite_vars(org_invite)
       new_user = !org_invite.user.confirmed?
-      
+
       {
         organization: org_invite.organization.name,
         team: org_invite.team ? org_invite.team.name : nil,
@@ -30,11 +30,11 @@ module MnoEnterprise
         invitee_email: org_invite.user.email,
       }
     end
-    
+
     describe 'confirmation_instructions' do
       describe 'new user' do
         it 'sends the right email' do
-          expect(MandrillClient).to receive(:deliver).with('confirmation-instructions',
+          expect(MnoEnterprise::MailClient).to receive(:deliver).with('confirmation-instructions',
             SystemNotificationMailer::DEFAULT_SENDER,
             { name: "#{user.name} #{user.surname}".strip, email: user.email },
             user_vars(user).merge(confirmation_link: routes.user_confirmation_url(confirmation_token: token))
@@ -49,7 +49,7 @@ module MnoEnterprise
         before { allow_any_instance_of(MnoEnterprise::User).to receive(:unconfirmed_email?).and_return(true) }
         
         it 'sends the right email' do
-          expect(MandrillClient).to receive(:deliver).with('reconfirmation-instructions',
+          expect(MnoEnterprise::MailClient).to receive(:deliver).with('reconfirmation-instructions',
             SystemNotificationMailer::DEFAULT_SENDER,
             { name: "#{user.name} #{user.surname}".strip, email: user.email },
             user_vars(user).merge(confirmation_link: routes.user_confirmation_url(confirmation_token: token))
@@ -62,7 +62,7 @@ module MnoEnterprise
     
     describe 'reset_password_instructions' do
       it 'sends the right email' do
-        expect(MandrillClient).to receive(:deliver).with('reset-password-instructions',
+        expect(MnoEnterprise::MailClient).to receive(:deliver).with('reset-password-instructions',
           SystemNotificationMailer::DEFAULT_SENDER,
           { name: "#{user.name} #{user.surname}".strip, email: user.email },
           user_vars(user).merge(reset_password_link: routes.edit_user_password_url(reset_password_token: token))
@@ -74,7 +74,7 @@ module MnoEnterprise
     
     describe 'unlock_instructions' do
       it 'sends the right email' do
-        expect(MandrillClient).to receive(:deliver).with('unlock-instructions',
+        expect(MnoEnterprise::MailClient).to receive(:deliver).with('unlock-instructions',
           SystemNotificationMailer::DEFAULT_SENDER,
           { name: "#{user.name} #{user.surname}".strip, email: user.email },
           user_vars(user).merge(unlock_link: routes.user_unlock_url(unlock_token: token))
@@ -90,7 +90,7 @@ module MnoEnterprise
       
       context 'when invitee is a confirmed user' do
         it 'sends the right email' do
-          expect(MandrillClient).to receive(:deliver).with('organization-invite-existing-user',
+          expect(MnoEnterprise::MailClient).to receive(:deliver).with('organization-invite-existing-user',
             SystemNotificationMailer::DEFAULT_SENDER,
             { name: "#{invitee.name} #{invitee.surname}".strip, email: invitee.email },
             invite_vars(org_invite).merge(confirmation_link: routes.org_invite_url(id: org_invite.id, token: org_invite.token))
@@ -104,7 +104,7 @@ module MnoEnterprise
         let(:invitee) { build(:user, :unconfirmed) }
         
         it 'sends the right email' do
-          expect(MandrillClient).to receive(:deliver).with('organization-invite-new-user',
+          expect(MnoEnterprise::MailClient).to receive(:deliver).with('organization-invite-new-user',
             SystemNotificationMailer::DEFAULT_SENDER,
             { email: invitee.email },
             invite_vars(org_invite).merge(confirmation_link: routes.user_confirmation_url(confirmation_token: invitee.confirmation_token))
@@ -118,7 +118,7 @@ module MnoEnterprise
 
     describe 'deletion_request_instructions' do
       it 'sends the correct email' do
-        expect(MandrillClient).to receive(:deliver).with('deletion-request-instructions',
+        expect(MnoEnterprise::MailClient).to receive(:deliver).with('deletion-request-instructions',
           SystemNotificationMailer::DEFAULT_SENDER,
           { name: "#{user.name} #{user.surname}".strip, email: user.email },
           user_vars(user).merge(terminate_account_link: routes.deletion_request_url(deletion_request))
@@ -130,7 +130,7 @@ module MnoEnterprise
 
     describe 'registration_instructions' do
       it 'sends the correct email' do
-        expect(MandrillClient).to receive(:deliver).with(
+        expect(MnoEnterprise::MailClient).to receive(:deliver).with(
             'registration-instructions',
             SystemNotificationMailer::DEFAULT_SENDER,
             { email: 'test@example.com' },